Health check against 0.0.0.0 instead of localhost
For the simulator & quoting service the HAPI.server is configured specifically to listen to 0.0.0.0. Other services don't set this and the default is 127.0.0.1.

Now recent changes to the docker engine made a change to how localhost is resolved, it used to be ipv4 but now is ipv6.

This means for people using a later version of docker the localhost would resolve to ::1 which the quoting service and simulator do not listen to.